[MODEXPS-70] - Update code to fix smells

+++ src/main/java/org/folio/des/scheduling/ExportTrigger.java
@@ -108,7 +108,7 @@ public class ExportTrigger implements Trigger {
       nextExecutionTime.setTime(lastActualExecutionTime);
       long diffHours = (nowDate.getTime() - nextExecutionTime.getTime().getTime())/(60 * 60 * 1000);
       if (diffHours > 0 && hours !=0 && diffHours > hours) {
-        int hoursToIncrease = Math.round(diffHours / hours) + 1;
+        int hoursToIncrease = (int)(Math.round((double) diffHours/hours) + 1);
         nextExecutionTime.add(Calendar.HOUR, hoursToIncrease * hours);
       } else {
         nextExecutionTime.add(Calendar.HOUR, hours);
